Fixed bug with attribute_data function
There was a bug in the attribute data function when merging a 1 km data grid directly with a mesh. The previous join function unordered the data during the merge, meaning the values were assigned to the wrong mesh nodes. This has been fixed by using the plyr::join function with a left hand join