Changed attribute_data function
The attribute data fucntion has been re-written to use the functions in the Raster package instead of manual averaging. The advantage of this is that there is less code to maintain. In addition, the function becomes more flexible with the type of gridded datasets that can be passed. It also avoids the need for manual dataframe merging (which can have issues)